db4

+++ src/main/resources/db/migration/V1__init3_schema.sql
@@ -2,14 +2,14 @@ CREATE EXTENSION IF NOT EXISTS btree_gist;
 CREATE EXTENSION IF NOT EXISTS pgcrypto;
 
 CREATE TABLE barber (
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   name TEXT NOT NULL,
   is_active BOOLEAN NOT NULL DEFAULT TRUE,
   created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
 
 CREATE TABLE service (
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   name TEXT NOT NULL,
   duration_min INT NOT NULL,
   buffer_after_min INT NOT NULL DEFAULT 0,
@@ -18,15 +18,15 @@ CREATE TABLE service (
 );
 
 CREATE TABLE client (
-  id SERIAL PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   name TEXT NOT NULL,
   phone TEXT,
   created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
 
 CREATE TABLE working_hours (
-  id SERIAL PRIMARY KEY,
-  barber_id INT NOT NULL REFERENCES barber(id) ON DELETE CASCADE,
+  id BIGSERIAL PRIMARY KEY,
+  barber_id BIGINT NOT NULL REFERENCES barber(id) ON DELETE CASCADE,
   day_of_week SMALLINT NOT NULL CHECK (day_of_week BETWEEN 1 AND 7),
   start_time TIME NOT NULL,
   end_time TIME NOT NULL,
@@ -35,8 +35,8 @@ CREATE TABLE working_hours (
 );
 
 CREATE TABLE time_off (
-  id SERIAL PRIMARY KEY,
-  barber_id INT NOT NULL REFERENCES barber(id) ON DELETE CASCADE,
+  id BIGSERIAL PRIMARY KEY,
+  barber_id BIGINT NOT NULL REFERENCES barber(id) ON DELETE CASCADE,
   starts_at TIMESTAMPTZ NOT NULL,
   ends_at TIMESTAMPTZ NOT NULL,
   reason TEXT,
@@ -45,9 +45,9 @@ CREATE TABLE time_off (
 
 CREATE TABLE appointment (
   id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-  barber_id INT NOT NULL REFERENCES barber(id),
-  service_id INT NOT NULL REFERENCES service(id),
-  client_id INT NOT NULL REFERENCES client(id),
+  barber_id BIGINT NOT NULL REFERENCES barber(id),
+  service_id BIGINT NOT NULL REFERENCES service(id),
+  client_id BIGINT NOT NULL REFERENCES client(id),
   starts_at TIMESTAMPTZ NOT NULL,
   ends_at TIMESTAMPTZ NOT NULL,
   status TEXT NOT NULL DEFAULT 'SCHEDULED',
@@ -66,4 +66,4 @@ ALTER TABLE appointment
   );
 
 CREATE INDEX IF NOT EXISTS idx_appointment_barber_start ON appointment (barber_id, starts_at);
-CREATE INDEX IF NOT EXISTS idx_timeoff_barber ON time_off (barber_id, starts_at);
+CREATE INDEX IF NOT EXISTS idx_timeoff_barber          ON time_off (barber_id, starts_at);
